Add files via upload
This is a script that scraps the URL of BBLAM (Bualuang Bank Asset Management).
I was creating this because I would like to reduce my burden's girlfriend she has already moved her job with a new role that contains an advice to the corporate fund at BBL.
Typically, I use the main package from Python Programming which is Selenium. It totally has a technique of the skillful with understanding the DOM-TREE and Regular Expression. However, I try to avoid of the second skill because my knowledge is limited. (my RE skill is always terrify)